add dtype

+++ cartpole/simulation/simulate.py
@@ -64,7 +64,7 @@ class Simulate(object):
             last_idx = self.agent.save_memory(state)
 
             # Chose action
-            inp = torch.from_numpy(np.array([self.agent.get_screen_history()])).type(torch.FloatTensor) / 255.0
+            inp = torch.from_numpy(np.array([self.agent.get_screen_history()])).type(torch.FloatTensor).type(utility.dtype) / 255.0
             action = self.agent.select_action(inp)
             # Action
             _, reward, done, _ = self.env.step(action)
